extract SlgContext into a struct
This struct encapsulates `max_size`
