+++ src/solve/slg/test.rs
@@ -5,6 +5,7 @@ use errors::*;
 use ir;
 use lower::*;
 use solve::slg::forest::Forest;
+use solve::slg::context::implementation::SlgContext;
 use std::sync::Arc;
 
 fn parse_and_lower_program(text: &str) -> Result<ir::Program> {
@@ -39,7 +40,7 @@ fn solve_goal(program_text: &str, goals: Vec<(usize, usize, &str, &str)>) {
             assert!(goal_text.ends_with("}"));
             let goal = parse_and_lower_goal(&program, &goal_text[1..goal_text.len() - 1]).unwrap();
             let peeled_goal = goal.into_peeled_goal();
-            let mut forest = Forest::new(env.clone(), max_size);
+            let mut forest = Forest::new(SlgContext::new(env, max_size));
             let result = format!("{:#?}", forest.force_answers(peeled_goal, num_answers));
 
             ::test_util::assert_test_result_eq(&expected, &result);
@@ -270,7 +271,7 @@ fn only_draw_so_many() {
     ir::tls::set_current_program(&program, || {
         let goal = parse_and_lower_goal(&program, goal_text).unwrap();
         let peeled_goal = goal.into_peeled_goal();
-        let mut forest = Forest::new(env.clone(), 10);
+        let mut forest = Forest::new(SlgContext::new(env, 10));
         let solution = forest.solve(&peeled_goal);
 
         // First, check we got the expected solution.
